requirements.txt
albumentations==1.0.3
pycocotools==2.0.2
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
seaborn>=0.11.0
catalyst==21.9
pytorch-toolbelt==0.4.4
segmentation-models-pytorch==0.2.0
scikit-learn==1.0